Restructured the Tests to address issue 12
