+++ src/main/config/teamengine/config.xml
@@ -13,11 +13,9 @@
         <local-name>ets-georss10-${version}</local-name>
         <title>ets-georss10 Conformance Test Suite</title>
         <description>Describe scope of testing.</description>
-        <!-- Add baisc conformance test suites
         <BasicConformanceClasses>
           <conformanceClass>Core</conformanceClass>
         </BasicConformanceClasses>
-        -->
       </suite>
       <revision>
         <name>${project.version}</name>
