--- src/main/java/org/opengis/cite/georss10/simple/package-info.java
@@ -1,14 +1,0 @@
-/**
- * Conformance Level 1 includes basic facilities. The following capabilities
- * must be supported: 
- * <ul>
- * <li>Capability 1.1</li>
- * <li>Capability 1.2</li>
- * <li>Capability 1.3</li> 
- * </ul>
- *
- * @see <a
- * href="http://www.w3.org/TR/html5/infrastructure.html#conformance-classes">HTML5
- * - Conformance classes</a>
- */
-package org.opengis.cite.georss10.simple;
